Rename dns directory to serializers
